ASB-123: Refactored existing registration journey for individuals into subfolders in preparation for separate feature files for organisation and agent
