+++ e2e-tests/cypress/integration/cookies-banner/cookies-banner.js
@@ -1,5 +1,5 @@
 import { Given, When, Then, And } from "cypress-cucumber-preprocessor/steps";
-import PO_WhoYouRegisterFor from '../who-are-you-registering-for/PageObjects/PO_WhoYouRegisterFor';
+import PO_WhoYouRegisterFor from '../registration/who-are-you-registering-for/PageObjects/PO_WhoYouRegisterFor';
 import PO_CookiePage from "./PageObjects/PO_CookiePage";
 const whoYouRegisterForPage = new PO_WhoYouRegisterFor();
 const cookiePage = new PO_CookiePage();
