+++ e2e-tests/cypress/integration/registration/myself/are-you-18-or-over/are-you-18-or-over.js
