+++ e2e-tests/cypress/integration/registration/myself/check-your-answers-before-registering/PageObjects/PO_CyaBeforeReg.js
