+++ e2e-tests/cypress/integration/registration/myself/check-your-answers-before-registering/check-your-answers-before-registering.js
