+++ e2e-tests/cypress/integration/registration/myself/full-name/PageObjects/PO_FullName.js
@@ -1,4 +1,4 @@
-import PageObject from "../../PageObject";
+import PageObject from "../../../../PageObject";
 
 class PO_FullName extends PageObject {
