+++ e2e-tests/cypress/integration/registration/myself/regsitration-complete/PageObjects/PO_RegComplete.js
