+++ e2e-tests/cypress/integration/registration/myself/regsitration-complete/regsitration-complete.js
