+++ e2e-tests/cypress/integration/registration/myself/uk-address-details/PageObjects/PO_AddressDetails.js
@@ -1,4 +1,4 @@
-import PageObject from "../../PageObject";
+import PageObject from "../../../../PageObject";
 
 class PO_AddressDetails extends PageObject{
