+++ e2e-tests/cypress/integration/registration/myself/uk-address-details/uk-address-details.js
