+++ e2e-tests/cypress/integration/registration/myself/what-do-you-want-to-tell-about-project/PageObjects/PO_TellAboutProject.js
@@ -1,4 +1,4 @@
-import PageObject from "../../PageObject";
+import PageObject from "../../../../PageObject";
 
 class PO_TellAboutProject extends PageObject{
