+++ e2e-tests/cypress/integration/registration/myself/what-is-your-email-address/PageObjects/PO_EmailAddress.js
@@ -1,4 +1,4 @@
-import PageObject from "../../PageObject";
+import PageObject from "../../../../PageObject";
 
 class PO_EmailAddress extends PageObject{
