+++ e2e-tests/cypress/integration/registration/myself/what-is-your-email-address/what-is-your-email-address.js
