+++ e2e-tests/cypress/integration/registration/myself/what-is-your-telephone-number.feature
@@ -1,4 +1,4 @@
-@testSuite
+@testSuite @registration @myself
 Feature: What is your telephone number page
   As a Test user
   I want to navigate to What is your telephone number page
