+++ e2e-tests/cypress/integration/registration/myself/what-is-your-telephone-number/PageObjects/PO_TeleNumber.js
@@ -1,4 +1,4 @@
-import PageObject from "../../PageObject";
+import PageObject from "../../../../PageObject";
 
 class PO_TeleNumber extends PageObject {
