+++ e2e-tests/cypress/integration/registration/myself/what-is-your-telephone-number/what-is-your-telephone-number.js
