+++ e2e-tests/cypress/integration/registration/who-are-you-registering-for.feature
@@ -1,4 +1,4 @@
-@testSuite
+@testSuite @registration @myself
 Feature: who are you registering for? page
     As a Test user
     I want to navigate to who are you registering for? page
