+++ e2e-tests/cypress/integration/registration/who-are-you-registering-for/PageObjects/PO_WhoYouRegisterFor.js
@@ -1,4 +1,6 @@
-class PO_WhoYouRegisterFor {
+import PageObject from "../../../PageObject";
+
+class PO_WhoYouRegisterFor extends PageObject {
 
     navigatetoTypeOfPartyPage() {
         cy.visit('/project-search', { failOnStatusCode: false });
