+++ e2e-tests/cypress/integration/registration/who-are-you-registering-for/who-are-you-registering-for.js
@@ -1,8 +1,6 @@
 import { Given, When, Then, And } from "cypress-cucumber-preprocessor/steps";
 import PO_WhoYouRegisterFor from './PageObjects/PO_WhoYouRegisterFor'
-import PageObject from '../PageObject'
 const whoYouRegisterForPage = new PO_WhoYouRegisterFor()
-const pageObject = new PageObject()
 
 Given('I navigate to the who are you registering for page', () => {
     whoYouRegisterForPage.navigatetoTypeOfPartyPage()
@@ -17,11 +15,11 @@ And("I can see the radio options content", () => {
 })
 
 And('I can see the logo gov uk text', () => {
-    pageObject.validateHeaderLogo()
+    whoYouRegisterForPage.validateHeaderLogo()
 })
 
 And('I can see the text This service is only for Application service', () => {
-    pageObject.validateHeaderContent()
+    whoYouRegisterForPage.validateHeaderContent()
 })
 
 Then('below error message should be presented on who are you registering for page', function (table) {
@@ -38,16 +36,16 @@ And('User clicks on continue button', () => {
 
 Then('I click on {string} logo', (logoLink) => {
     switch (logoLink) {
-        case "planning inspectorate": pageObject.clickOnPlanningInspectorateLogo();
+        case "planning inspectorate": whoYouRegisterForPage.clickOnPlanningInspectorateLogo();
             break;
-        case "crown copyright": pageObject.clickOnCrownCopyRight();
+        case "crown copyright": whoYouRegisterForPage.clickOnCrownCopyRight();
             break;
         default: throw console.error('uanble to find specified logo link: ' + logoLink);
     }
 })
 
 Then('I click on feedback link', () => {
-    pageObject.clickOnProvideFeedbackLink()
+  whoYouRegisterForPage.clickOnProvideFeedbackLink()
 })
 
 Then('I click on back link', () => {
